+++ test/usecases/user_usecases_test.dart
@@ -148,6 +148,12 @@ class MockUserRepository implements UserRepository {
   }
 
   @override
+  Future<Either<Failure, bool>> checkUserFollow(String username) => throw UnimplementedError();
+  @override
+  Future<Either<Failure, void>> followUser(String username) => throw UnimplementedError();
+  @override
+  Future<Either<Failure, void>> unfollowUser(String username) => throw UnimplementedError();
+  @override
   Future<Either<Failure, void>> deleteGPGKey(int id) async {
     if (shouldFail) return Left(failure);
     return Right(null);